rating int->float fix

+++ api/EatAndWatch/Entities/DTO/MovieDto.cs
@@ -6,7 +6,7 @@ namespace Entities.DTO
     {
         public string Name { get; set; }
         public string? Description { get; set; }
-        public int Rating { get; set; } = 0;
+        public float Rating { get; set; } = 0;
         public IEnumerable<int> GenreIds { get; set; } = [];
         public IEnumerable<int> TagIds { get; set; } = [];
         public string? Image { get; set; }
